requirements.txt
torch==2.0.0
numpy==1.24.0
python-chess==1.99
pygame==2.2.0
matplotlib==3.7.0
stockfish==3.28.0
requests==2.31.0
trueskill==0.4.5